+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+* Add AuthorizationAdjustment class and `authorizationAdjustments` to Transaction
 * Add iDEAL webhook support
 * Add `IDEAL_PAYMENT` to `PaymentInstrumentType`
 * Create Braintree class to be PSR compliant